Implements clear field functionality
Enables the user to clear the game field, resetting the values of all non-predefined cells.

This commit adds a `clearField` method to the `Game` struct that resets the value of each cell while preserving predefined values. The `GameView` is updated to call this function, allowing the user to reset the field.